Populate test-server.json from example.
I figured there was no reason not to do this the same as the client.

I recommend re-cloning and re-running setup_dev.sh after this commit
so that nothing gets confused.

Also fixed a bug where sometimes the old instance of the server
would be running when you tried to test.